+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.11 branch
 ### 2.11.0
+* requires at least PHP 7.4;
+* added `MySql::getAuthFile()` method. So the `MySql::$auth` property is now private;
 * `createBackup()` and `createSomeBackups()` are now testing global functions and no longer methods provided by the
   `TestCase` class;
 * added `CommandTestCase` to test commands.